Added the foundations for adding parts to the player's inventory (all managed within MechanicScript) but still unsure how to do a UI click event.
Turned the MechanicScript into a persisting singleton so that it can store all data relating to parts and persist them across scenes. (Might rename to PartManager)

Stores List of Parts that serves as the player's inventory. Has functions to add to this list and compute total stats based on this list. PlayerController gets this data on Start.